Reorganized Files into Areas
Created two Areas within the application: "Admin" and "Customer"; moved Category Controller and Views into Admin Area, and relocated the Home Controller and Views into the Customer Area. NOTE: Although explicitly defining the Area in the Controller files is no longer needed in .NET 6; this was done as a safety measure to ensure there are no issues with this implementation.